+++ src/keymap/getkey.h
@@ -13,7 +13,7 @@ Copyright 2016 Stichting Humanity4all
 
 namespace keymap_n {
 
-translation_service_n::key_t get_key(int layer, int switch1, int switch2);
+translation_service_n::key_t get_key(uint8_t layer, uint8_t switch1, uint8_t switch2);
 
 } // namespace keymap_n
